Add shared object support
This feature is highly experimental yet and must be tested. But, the
proof-of-concept works.
For now, the persistency of shared objects is partially supported and is
done in RAM only. So all persistent shared objects are lost when nginx is
restarted. When a request was made for an existing shared object, persitency
flags are ignored (we keep flags of the request which creates the shared
object), so "SharedObject.BadPersistence" error is never thrown.

4 netStatus events were created to be dispatched to clients:

o SharedObject.Connect.Success: The server sends this event to notify the
                                client on a successful connection.

o SharedObject.Connect.Closed:  The server sends this event to client in
                                response to "Release" event.

o SharedObject.Peer.Connect:    The server sends this event to notify all
                                clients, except the client originating
                                the request, on a successful connection.

o SharedObject.Peer.Disonnect:  The server sends this event to notify all
                                clients, except the client originating
                                the request, on a disconnection.